pyproject.toml
[build-system]
requires = [
  "setuptools>=62.1",
  "setuptools_scm[toml]>=6.2",
  "wheel",
]
build-backend = "setuptools.build_meta"

[project]
name = "aiapy"
description = "Python library for AIA data analysis."
requires-python = ">=3.10"
readme = { file = "README.rst", content-type = "text/x-rst" }
license = { file = "licenses/LICENSE.rst" }
authors = [
  { name = "AIA Instrument Team @ LMSAL", email = "nfreij@set.org" },
]
dependencies = [
  'sunpy[net,image,map]>=6.0',
]
dynamic = ["version"]
keywords = [
  "solar physics",
  "solar",
  "science",
  "NASA",
  "SDO",
  "AIA",
  "solar dynamics observatory",
  "atmospheric imaging assembly",
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: BSD License",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Astronomy",
]

[project.optional-dependencies]
all = ["aiapy"]
tests = [
  "pytest",
  "pytest-doctestplus",
  "pytest-cov",
  "pytest-xdist",
  "aiapy[all]",
  "hissw",
  "pytest-astropy",
]
docs = [
  "sphinx",
  "sphinx-automodapi",
  "sphinx-changelog",
  "packaging",
  "aiapy[all]",
  "sphinx-copybutton",
  "sphinx-design",
  "sphinx-gallery",
  "sphinx-hoverxref",
  "sphinxext-opengraph",
  "sunpy-sphinx-theme",
  "sunpy-sphinx-theme",
]

[project.urls]
Homepage = "https://aia.lmsal.com/"
"Source Code" = "https://github.com/LM-SAL/aiapy"
Download = "https://pypi.org/project/aiapy"
Documentation = "https://aiapy.readthedocs.io/en/stable/"
Changelog = "https://aiapy.readthedocs.io/en/stable/changelog.html"
"Issue Tracker" = "https://github.com/LM-SAL/aiapy/issues/"

[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
include = ["aiapy*"]
exclude = ["aiapy._dev*"]

[tool.setuptools_scm]
write_to = "aiapy/_version.py"

[tool.gilesbot]
  [tool.gilesbot.pull_requests]
  enabled = true

  [tool.gilesbot.towncrier_changelog]
  enabled = true
  verify_pr_number = true
  changelog_skip_label = "No Changelog Entry Needed"
  help_url = "https://github.com/LM-SAL/aiapy/blob/main/changelog/README.rst"

  changelog_missing_long = "There isn't a changelog file in this pull request. Please add a changelog file to the `changelog/` directory following the instructions in the changelog [README](https://github.com/LM-SAL/aiapy/blob/main/changelog/README.rst)."

  type_incorrect_long = "The changelog file you added is not one of the allowed types. Please use one of the types described in the changelog [README](https://github.com/LM-SAL/aiapy/blob/main/changelog/README.rst)"

  number_incorrect_long = "The number in the changelog file you added does not match the number of this pull request. Please rename the file."

# TODO: This should be in towncrier.toml but Giles currently only works looks in
# pyproject.toml we should move this back when it's fixed.
[tool.towncrier]
  package = "aiapy"
  filename = "CHANGELOG.rst"
  directory = "changelog/"
  issue_format = "`#{issue} <https://github.com/LM-SAL/aiapy/pull/{issue}>`__"
  title_format = "{version} ({project_date})"

  [[tool.towncrier.type]]
  directory = "breaking"
  name = "Breaking Changes"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "deprecation"
  name = "Deprecations"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "removal"
  name = "Removals"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "feature"
  name = "New Features"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "bugfix"
  name = "Bug Fixes"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "doc"
  name = "Documentation"
  showcontent = true

  [[tool.towncrier.type]]
  directory = "trivial"
  name = "Internal Changes"
  showcontent = true
